rewrite device memory system to dx12 & vulkan like architecture

+++ Engine/include/Engine/Runtime/RHI/Exar/ExarBufferTypes.hpp
@@ -8,7 +8,7 @@
 namespace UnasciiEngine::RHI::EXAR
 {
 	struct ExarBufferDesc {
-		u64 size;
+		size_t size;
 		ExarUsage usage;
 		ExarBindFlag bindFlags;
 	};
